+++ paper/paper.bib
@@ -5,7 +5,7 @@ volume = {163},
 pages = {105679},
 year = {2023},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2023.105679},
+doi = {10.1016/j.envsoft.2023.105679},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815223000658},
 author = {A. Reyes and G. Viera-López and J.J. Morgado-Vega and E. Altshuler},
 keywords = {Trajectory analysis, Modeling, Tracking, Python}
@@ -16,6 +16,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   author    = {Zheng, Yu and Zhang, Lizhu and Xie, Xing and Ma, Wei-Ying},
   booktitle = {Proceedings of the 18th international conference on World wide web},
   pages     = {791--800},
+  doi       = {10.1145/1526709.1526816},
   year      = {2009}
 }
 
@@ -24,6 +25,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   author    = {Zheng, Yu and Li, Quannan and Chen, Yukun and Xie, Xing and Ma, Wei-Ying},
   booktitle = {Proceedings of the 10th international conference on Ubiquitous computing},
   pages     = {312--321},
+  doi       = {10.1145/1409635.1409677},  
   year      = {2008}
 }
 
@@ -60,6 +62,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   volume  = {141},
   number  = {10},
   pages   = {3576--3592},
+  doi     = {10.1175/mwr-d-12-00254.1},
   year    = {2013}
 }
 
@@ -70,6 +73,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   volume  = {31},
   number  = {2},
   pages   = {287--301},
+  doi     = {10.1175/JTECH-D-12-00119.1},
   year    = {2014}
 }
 
@@ -81,6 +85,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   number    = {4},
   pages     = {690--699},
   year      = {2021},
+  doi       = {10.1007/s00376-020-0211-7},
   publisher = {Springer}
 }
 
@@ -90,6 +95,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   booktitle    = {2019 8th Brazilian Conference on Intelligent Systems (BRACIS)},
   pages        = {788--793},
   year         = {2019},
+  doi          = {10.1109/bracis.2019.00141},
   organization = {IEEE}
 }
 
@@ -101,6 +107,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   number    = {2},
   pages     = {57},
   year      = {2017},
+  doi       = {10.3390/ijgi6020057},
   publisher = {MDPI}
 }
 
@@ -110,6 +117,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   booktitle    = {2022 IEEE International Conference on Big Data and Smart Computing (BigComp)},
   pages        = {331--333},
   year         = {2022},
+  doi          = {10.1109/bigcomp54360.2022.00074},
   organization = {IEEE}
 }
 
@@ -121,6 +129,7 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   number    = {3},
   pages     = {032402},
   year      = {2020},
+  doi       = {10.1103/physreve.102.032402},
   publisher = {APS}
 }
 
@@ -132,5 +141,6 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   number    = {21},
   pages     = {10979},
[…]
