### 2023-06-11 #5543 @gvieralopez
Dear @arfon ,

I sincerely appreciate your prompt response and assistance. I have updated all the missing DOIs in the references, and I can confirm that all the suggestions provided by the bot were accurate.

I would like to bring to your attention that part of the code for the functioning of pactus is hosted in a complementary repository, which we have referenced in the documentation: https://github.com/yupidevs/trajectory-datasets. This repository contains the code that manages the backend of the library, specifically for the standardization of data for each compatible dataset and hosting the datasets.

It is worth noting that we made a conscious decision to keep the code of the library simple and extensible by maintaining all dataset-specific code independent from the framework. Consequently, we did not include this particular code in the library repository. Although this may give the impression that the project code base is smaller, our team main focus is on delivering a library that is both functional and scalable for the research community.

Thank you for your time and support.

Best regards,

Gustavo (In the name of all authors)
