### 2023-07-12 #5543 @Kevin-Mattheus-Moerman
@gvieralopez I've removed the query-scope label as several members of the editorial board felt like this may be in scope for JOSS. As you can see :point_up: we've started looking for editors. However, many are handling other papers already or are currently taking annual leave. Hence I may have to label this paper as `waitlisted` until an editor becomes available. Thanks for your patience.
