### 2023-07-17 #5543 @gvieralopez
Dear @arfon

After reviewing the list, I have identified some candidates with relevant backgrounds that may be suitable for the review process.

For expertise in time series analysis and classification, you may consider vjbytes102, daviddewhurst, ulf1, and jwwthu. Although their fields may differ from trajectory classification, their experience in general time series analysis should bring valuable insights to the paper.

Regarding data preprocessing and reproducible research, I suggest considering adrien-matta, hseabolt, tkmamidi, and dvanic. They could offer constructive feedback on how to improve the software and the paper to make it more valuable and compliant with open data research best practices.

Finally, for fields that may benefit from the content of the paper, I recommend contacting jedalong (Movement ecology), felixhenninger (behavioral science/movement tracking), and nsaunier (transportation). Their expertise in these fields could help ensure the paper's relevance and significance.

I hope this information is helpful, and please let me know if you need any further assistance.

Best regards,

Gustavo
