### 2023-07-23 #5543 @arfon
:wave: @vjbytes102 @ulf1 @adrien-matta @tkmamidi @jedalong @felixhenninger would any of you be willing to review this submission for JOSS? The submission under consideration is _pactus: A Python framework for trajectory classification_.

Many thanks
Arfon
