### 2023-07-31 #5543 @arfon
@JustinShenk @laempy – would you be willing to review this submission for JOSS? The submission under consideration is _pactus: A Python framework for trajectory classification_.

Many thanks
Arfon
