### 2023-08-01 #5543 @arfon
@matt002 @finsberg @abhishektha – would you be willing to review this submission for JOSS? The submission under consideration is pactus: A Python framework for trajectory classification.

Many thanks
Arfon
