### 2023-08-04 #5543 @arfon
:wave: @miladmzdh – would you be willing to review this submission for JOSS? The submission under consideration is pactus: A Python framework for trajectory classification.

Many thanks
Arfon
