### 2023-08-04 #5738 @editorialbot
[REVIEW]: pactus: A Python framework for trajectory classification
**Submitting author:** @gvieralopez (<a href="http://orcid.org/0000-0002-9661-5709">Gustavo Viera-López</a>)
**Repository:** https://github.com/yupidevs/pactus
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.0
