### 2023-08-25 #5738 @miladmzdh
Hello @gvieralopez,

Impressive effort. I appreciate the structured code and thorough documentation; it's evident that a lot of care went into this. Reviewing your code was a pleasant experience. However, I do have a couple of concerns that I'd like to address:

- When I installed the package using "pip install pactus" it could not locate any of the datasets. However, when I pulled it and installed it locally, it worked completely fine. Maybe it's my machine, but maybe it's something to look into?

- While the datasets used here are well-known, it is a good practice to add an example to your examples, even something simple like this:

from yupi.graphics import plot_2d
plot_2d(Dataset.animals().trajs[0], legend=None)

So users can explore the trajectories better and have a sense of what they are working with (both visually and the data attributes). I know your package is not about the exploration of the data but when we have several datasets, better to have some tools to explore or at least have examples to explore the data. You might even have a completely separate example code to explore the datasets a bit.

- While readers and users might be experts in the field and are aware that the numbers might be different than yours, I would add a sentence in the "getting started" section and mention that they should not expect the same numbers. (I understand you mentioned "similar to", but maybe more explicitly.)

- And finally, as Justin mentioned, there is no mention of other packages, what they lack, and how pactus improves those.

Good luck and let me know when these issues are taken care of so I can check those points in my "review checklist".
