### 2023-08-27 #5738 @gvieralopez
@JustinShenk @miladmzdh thank you very much for the fast and constructive feedback.

@JustinShenk
We just started a thorough review of all the dependencies and versions. Thank you for noticing the issue with the incorrect version of numpy in the requirements.txt. I will write another message in this issue once we update both, paper and code, so you can check them together.

@miladmzdh
We appreciate your feedback and your positive comments.
We will add an example showcasing all the data exploration that an user can do with a dataset. That is a Great idea.
Regarding the reproducibility of the examples, we will be more explicit on the stochastic variations that can occur if  RNG are used with different seeds.
We will update the manuscript with a more comprehensive review of the available software.
Finally, we are very interested on detecting and fixing the source of the error you mentioned during the installation. Could you please open an issue in the GitHub repository describing the exact error and system details, so we can reproduce it.

Best,

Gustavo, in the name of all authors.
