### 2023-08-28 3c43470
Update dependencies versions
Co-authored-by: gvieralopez <gvieralopez@gmail.com>
