+++ pyproject.toml
@@ -19,9 +19,9 @@ keywords = [
     "classification"
 ]
 dependencies = [
-    "numpy >= 1.20.0",
+    "numpy >= 1.23.5",
     "yupi >= 0.11.2",
-    "tensorflow >= 2.9.1",
+    "tensorflow >= 2.12.0",
     "scikit-learn >= 1.1.1",
     "xgboost >= 1.7.4",
     "GitPython >= 3.1.29"
