### 2023-08-28 e4ea76b
Update readme example
Co-authored-by: gvieralopez <gvieralopez@gmail.com>
