+++ .readthedocs.yaml
@@ -2,6 +2,11 @@
 # Required
 version: 2
 
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    builder: html
@@ -11,8 +16,6 @@ sphinx:
 formats:
    - pdf
 
-# Optionally set the version of Python and requirements required to build your docs
 python:
-   version: 3.7
    install:
      - requirements: docs/requirements.txt
