+++ paper/paper.bib
@@ -144,3 +144,29 @@ keywords = {Trajectory analysis, Modeling, Tracking, Python}
   doi       = {10.3390/app122110979},
   publisher = {Multidisciplinary Digital Publishing Institute}
 }
+
+@article{shenk2021traja, 
+  title = {Traja: A Python toolbox for animal trajectory analysis}, 
+  author = {Shenk, Justin and Byttner, Wolf and Nambusubramaniyan, Saranraj and Zoeller, Alexander},
+  journal = {Journal of Open Source Software},
+  volume = {6}, 
+  number = {63}, 
+  pages = {3202},  
+  year = {2021}, 
+  doi = {10.21105/joss.03202},
+  url = {https://doi.org/10.21105/joss.03202},
+  publisher = {The Open Journal}
+} 
+
+@article{joo2020navigating,
+  title={Navigating through the R packages for movement},
+  author={Joo, Rocio and Boone, Matthew E and Clay, Thomas A and Patrick, Samantha C and Clusella-Trullas, Susana and Basille, Mathieu},
+  journal={Journal of Animal Ecology},
+  volume={89},
+  number={1},
+  pages={248--267},
+  year={2020},
+  doi = {10.1111/1365-2656.13116},
+  url = {https://doi.org/10.1111/1365-2656.13116},
+  publisher={Wiley Online Library}
+}
