+++ paper/paper.md
@@ -19,7 +19,7 @@ authors:
     orcid: 0000-0001-7305-4710
     equal-contrib: true
     affiliation: 3
-  - name: E. Altshuler
+  - name: E.~Altshuler
     orcid: 0000-0003-4192-5635
     equal-contrib: false
     affiliation: 3
@@ -55,7 +55,8 @@ been studied in many different fields, including robotics, behavior analysis,
 mobility pattern mining, and user activity recognition [@da2019survey]. This task
 presents multiple challenges for conventional classification models, such as the 
 indeterminate length of trajectories [@li2022two], the range of entities that generate 
-trajectories [@janczura2020classification; @xiao2017identifying; @bae2022transformer], and the absence of established standards in trajectory datasets 
+trajectories [@janczura2020classification; @xiao2017identifying; @bae2022transformer], 
+and the absence of established standards in trajectory datasets 
 [@xiao2017identifying; @bae2022transformer]. 
 
 Our study endeavors to lay the foundation for the assessment of innovative 
@@ -64,9 +65,25 @@ a new framework, referred to as `pactus`, which addresses the challenges of
 trajectory classification by providing direct access to a carefully chosen 
 collection of datasets and several trajectory classifiers. `pactus` 
 facilitates researchers' ability to experiment with various approaches 
-and assess their performance on different types of data.
+and assess their performance on different types of data. A comprehensive software 
+documentation is provided on 
+([https://pactus.readthedocs.io/en/latest/](pactus.readthedocs.io)).
 
 # Statement of need
+
+In recent years, several software libraries have emerged, aiming to automate trajectory 
+data analysis. Within the R community, there are various available tools 
+[@joo2020navigating]. Recognizing the popularity and extensive usage of Python, the `traja`
+software [@shenk2021traja] was developed to integrate different analysis techniques for 
+two-dimensional trajectories, primarily focusing on animal behavioral analysis. Additionally, 
+the `yupi` library [@reyes2023yupi] was created to handle trajectory analysis for applications 
+involving an arbitrary number of dimensions.
+
+Although these libraries offer valuable tools for trajectory classification, such as 
+classification models and feature extraction from trajectories, they were not specifically 
+designed for this task. Consequently, contemporary research on trajectory classification
+faces limitations in terms of evaluation, often considering only a limited number of datasets 
+or reporting only a reduced set of metrics [@bae2022transformer].
 
 The lack of standardization in trajectory datasets, coupled with the difficulty 
 of obtaining these datasets for evaluation, poses a significant challenge to 
@@ -88,18 +105,18 @@ researchers to distribute their findings as simple Python scripts, relying on `p
 for all tasks related to data acquisition, processing, and model evaluation.
 
 
-# Software Overview
+# Pactus Software Library
 
-The functionalities of `pactus` can be divided into four different categories as shown in
-\autoref{fig:overview}.
+The functionalities of `pactus` can be divided into modules: Data handling, Feature extraction,
+Classification models and Evaluation.
 
-![Overview of the resources available in `pactus` coupled with an usage example.\label{fig:overview}](1.pdf)
+## Data handling
 
-
-The selection of datasets was conducted with meticulous care to encompass a broad 
+The library provides direct access to some of the most commonly used datasets for trajectory 
+classification. The selection of datasets was conducted with meticulous care to encompass a broad 
 range of trajectories and classification objectives. Our initial selection includes
 GeoLife [@zheng2009mining; @zheng2008understanding; @zheng2010geolife], The Starkey 
-Project dataset, also known as `Animals' in the	trajectory classification 
+Project dataset, also known as `Animals` in the	trajectory classification 
 community [@rapp2009elk], four different datasets from the the UCI repository 
 [@Dua:2019] and two different hurricane datasets, provided by National Hurricane 
[…]
