### 2023-08-30 #5738 @gvieralopez
Dear @JustinShenk,

Following your suggestions, we have made the following updates to the project:

1. We reviewed and updated the versions of some of our dependencies. We tested the installation on several systems, and the software should now be much more stable. We thank you once again for spotting the issue and pointing us in the direction of the solution with your pull request. The release of pactus 0.4.1 contains this fix.
2. We have added additional references to related software for trajectory analysis in the paper. These references provide useful tools for trajectory classification.
3. We have subdivided the section of the paper that explains the software's structure into subsections. This division allows us to highlight important features from each category more effectively.

We appreciate all the effort you put into your response and recommendations, and we are very grateful for your review.
