### 2023-08-30 #5738 @gvieralopez
Dear @miladmzdh,

According to your recommendations, we have made the following changes to the project:

1. We have included an additional example in the software documentation that demonstrates data exploration on the available datasets in Pactus. You can check out the example at: https://pactus.readthedocs.io/en/latest/examples/example0.html. ↗️

2. Regarding the variabilities that can occur due to stochastic elements in the process, Pactus allows for setting a seed for the RNG used at each stage. However, we noticed that the example in the readme did not use a fixed seed. As a result, the evaluation results might differ slightly from the ones shown in the readme itself. To address this, we have added extra code to fix the seed in the example from the readme and provided a brief explanation for its necessity.

3. We have updated the paper with additional references to related software for trajectory analysis that also provide useful tools for trajectory classification.

4. We have divided the section of the paper that explains the software's structure into subsections. This division allows us to highlight important details more effectively.

Regarding the installation issue you reported, we have been unable to reproduce it. If possible, could you provide a detailed description of the exact error and your system details? You can open an issue in the software repository, and we will ensure to find a solution.

We would like to express our gratitude for the time you dedicated to reviewing our work and for providing suggestions to improve it.
