+++ docs/source/conf.py
@@ -17,7 +17,7 @@ sys.path.insert(0, os.path.abspath("../../pactus/models/"))
 project = "pactus"
 copyright = "2023, Jorge Morgado Vega"
 author = "Jorge Morgado Vega, Gustavo Viera-López, Alfredo Reyes"
-release = "0.4.1"
+release = "0.4.2"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
