+++ pactus/__init__.py
@@ -6,7 +6,7 @@ from yupi.core import featurizers
 from pactus.dataset import Dataset
 from pactus.models import Evaluation, EvaluationComparison
 
-__version__ = "0.4.1"
+__version__ = "0.4.2"
 
 __all__ = [
     "Dataset",
