### 2023-09-17 #5738 @gvieralopez
Dear @arfon,

We deeply appreciate your guidance throughout the revision of our manuscript. We have reviewed the paper one more time and all appears to be correct. Furthermore, a new version of the software, which integrates all modifications issued from this review, has been released.

We have prepared and uploaded an archive of this software on Zenodo. The archive mirrors the title and authors of the JOSS paper, maintaining consistency. The archive is accessible via the following link: https://zenodo.org/record/8352324. The corresponding DOI is: **10.5281/zenodo.8352324**.

We are thankful for your assistance, and the  insights provided during this process by all the reviewers. We look forward to any further steps that may be necessary.

Best Regards,

Gustavo, on behalf of all the authors
