### 2023-09-22 #5738 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
